apps/stream-relay/requirements.txt
websockets>=12.0
aiohttp>=3.9
Pillow>=10.0
